Refactor `strip_softmax`
turns out this actually wasn't necessary as `softmax` cannot be used as an activation function inside `Dense`:

```
`softmax(x)` called with a number, but it expects an array. Usually this is because a layer like `Dense(3,4,softmax)` is broadcasting it like an activation function; `softmax` needs to be outside the layer.
```